+++ rest_framework_pydantic_ai/types/__init__.py
@@ -1,3 +1,4 @@
 from rest_framework_pydantic_ai.types.agent_deps import AgentDeps
+from rest_framework_pydantic_ai.types.query_param import QueryParam
 
-__all__ = ["AgentDeps"]
+__all__ = ["AgentDeps", "QueryParam"]